Adicionando dockerfile e erros

+++ backend/services/expense_service.py
@@ -41,12 +41,17 @@ async def delete_expense(
         db: Session,
         user: user_schema.UserSchema,
         expense_id: int):
-    print(await get_transactions(db, user, {"category_expense_id": expense_id}))
     transactions = await get_transactions(db, user, {"category_expense_id": expense_id})
     if transactions:
         raise fastapi.HTTPException(
-            status_code=400,
+            status_code=422,
             detail="This expense is in use and cannot be deleted")
+    
+    expense = await expense_repository.get_expense_by_id(db, user, expense_id) 
+    if expense is None:
+        raise fastapi.HTTPException(
+            status_code=404, detail="Expense not found")
+    
     return await expense_repository.delete_expense(db, user, expense_id)
 
 
@@ -56,4 +61,11 @@ async def update_expense(
     expense: expense_schema.ExpenseCreateSchema,
     expense_id: int,
 ):
+    
+    new_expense = await expense_repository.get_expense_by_id(db, user, expense_id)
+
+    if new_expense is None:
+        raise fastapi.HTTPException(
+            status_code=404, detail="Expense not found")
+
     return await expense_repository.update_expense(db, user, expense, expense_id)
